Update index.html
Hey! For Bayes theorem, part 3: Bayes with the General Law of Probability, shouldn't the last equations denominator use j = 1, instead of i (This is what I changed). 
Since i is used in the Bayes equations numerator and this makes it seem that the Sum notation of the denominator is also dependent of the changing i's.